Docs nomenclature correction
