+++ doce_conta/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:doce_conta/view/Initial/Initial_screen.dart';
+import 'package:doce_conta/view/Initial/initial_screen.dart';
 void main() {
   runApp(const MyApp());
 }
