+++ doce_conta/lib/view/Initial/register_screen.dart
@@ -1,5 +1,5 @@
 import 'package:doce_conta/view/Initial/login_screen.dart';
-import 'package:doce_conta/view/Initial/successful_register.dart';
+import 'package:doce_conta/view/Initial/successful_register_screen.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import '../../widgets/input_and_label.dart';
