+++ doce_conta/lib/view/Product/new_product_screen.dart
@@ -1,4 +1,4 @@
-import 'package:doce_conta/view/Product/concluded.dart';
+import 'package:doce_conta/view/Product/concluded_screen.dart';
 import 'package:doce_conta/view/Product/icon_selection_screen.dart';
 import 'package:doce_conta/view/Product/profit_margin_screen.dart';
 import 'package:flutter/material.dart';
