+++ doce_conta/lib/view/Reports/hub_report_generation_screen.dart
@@ -1,4 +1,4 @@
-import 'package:doce_conta/view/Reports/annual_report.dart';
+import 'package:doce_conta/view/Reports/annual_report_screen.dart';
 import 'package:doce_conta/widgets/container_custom_button.dart';
 import 'package:flutter/material.dart';
