+++ doce_conta/lib/view/Reports/monthly_report_screen.dart
@@ -1,91 +1,84 @@
-import 'package:doce_conta/view/Reports/report_descriptions_screen.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
+import 'package:flutter_svg/svg.dart';
+import '../../widgets/conteiner_text_center.dart';
 
-class MontlhyReport extends StatelessWidget {
+class MonthReport extends StatelessWidget {
+  const MonthReport({super.key});
+
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Color(0xFF98CFC2),
-      appBar: AppBar(
-        backgroundColor: const Color(0xFF00382B),
-        leading: IconButton(
-          icon: const Icon(Icons.arrow_back, color: Colors.white),
-          onPressed: () {
-            Navigator.of(context).pop();
-          },
+    return MaterialApp(
+      home: Scaffold(
+        backgroundColor: Color(0xFFE1FAF4),
+        appBar: AppBar(
+          leading:IconButton(
+            icon:const Icon(Icons.arrow_back, color: Colors.white),
+            onPressed: (){
+              Navigator.pop(context);
+            }),
+          centerTitle: true,
+          title: SvgPicture.asset(
+              "assets/Images/relatorio.svg",
+              height: 40,
+              color: const Color(0xFFF5A001),
+            ),
+          backgroundColor: const Color(0xFF00382B),
         ),
-        title: SvgPicture.asset(
-          'assets/images/relatorio.svg',
-          height: 40,
-          color: const Color(0xFFF5A001),
-        ),
-        centerTitle: true,
-      ),
-      body: Column(
-        children: [
-          // Lista de produtos
-          Expanded(
-            child: ListView.separated(
-              padding: const EdgeInsets.all(16.0),
-              itemCount: 5,
-              itemBuilder: (context, index) {
-                // Lista de produtos
-                final infoName = ['Quantidade de itens vendidos', 'Valor bruto: caixa', 'Lucro diário', 'Valor Contas', 'Reserva'];
-                final infoPrice = ['x', 'R\$', 'R\$', 'R\$', 'R\$'];
-                return _buildInfoItem(infoName[index], infoPrice[index]);
-              },
-              separatorBuilder: (context, index) => Divider(
-                color: Colors.amber, // Linha amarela entre os itens
-                thickness: 1,
-                height: 16, // Ajusta o espaçamento da linha entre os itens
+        body: Center(
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Conteinertextcenter(btnText: "Janeiro", widhtContainer: 151.54, heightContainer: 81.2),
+                  SizedBox(width: 20),
+                  Conteinertextcenter(btnText: "Fevereiro", widhtContainer: 151.54, heightContainer: 81.2),
+                ],
               ),
-            ),
+              Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Conteinertextcenter(btnText: "Março", widhtContainer: 151.54, heightContainer: 81.2),
+                  SizedBox(width: 20),
+                  Conteinertextcenter(btnText: "Abril", widhtContainer: 151.54, heightContainer: 81.2),
+                ],
+              ),
+              Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Conteinertextcenter(btnText: "Maio", widhtContainer: 151.54, heightContainer: 81.2),
+                  SizedBox(width: 20),
+                  Conteinertextcenter(btnText: "Junho", widhtContainer: 151.54, heightContainer: 81.2),
+                ],
+              ),
+              Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Conteinertextcenter(btnText: "Julho", widhtContainer: 151.54, heightContainer: 81.2),
+                  SizedBox(width: 20),
+                  Conteinertextcenter(btnText: "Agosto", widhtContainer: 151.54, heightContainer: 81.2),
+                ],
+              ),
+              Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Conteinertextcenter(btnText: "Setembro", widhtContainer: 151.54, heightContainer: 81.2),
+                  SizedBox(width: 20),
+                  Conteinertextcenter(btnText: "Outubro", widhtContainer: 151.54, heightContainer: 81.2),
+                ],
+              ),
+              Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Conteinertextcenter(btnText: "Novembro", widhtContainer: 151.54, heightContainer: 81.2),
+                  SizedBox(width: 20),
+                  Conteinertextcenter(btnText: "Dezembro", widhtContainer: 151.54, heightContainer: 81.2),
+                ],
+              ),
+            ],
           ),
-          // Botão de adicionar
-          Padding(
-            padding: const EdgeInsets.only(bottom: 16.0),
-            child: ElevatedButton(
-              onPressed: () {
-                //Linkar página de descrições
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) => ReportDescriptions()));
-              },
-              child: 
-              const Text(
-                "Descrições",
-                style: TextStyle(color: Colors.black),
-              )
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
- Widget _buildInfoItem(String infoName, String infoPrice) {
-    return Container(
-      margin: const EdgeInsets.symmetric(vertical: 8.0),
-      padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 12.0),
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.circular(25.0),
-        border: Border.all(color: Colors.grey, width: 1),
-      ),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Text(
-            infoName,
-            style: const TextStyle(fontSize: 16, color: Colors.black),
-          ),
-          Text(
-            infoPrice,
-            style: const TextStyle(fontSize: 16, color: Colors.black),
-          ),
-        ]
+        )
       ),
     );
   }
