+++ doce_conta/lib/view/Stock/controll_by_values_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:doce_conta/view/Stock/graphics.dart';
+import 'package:doce_conta/view/Stock/dashboard_screen.dart';
 
 class ValueControlScreen extends StatelessWidget {
   @override
