+++ doce_conta/lib/widgets/container_custom_button.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
-import '../widgets/button.dart';
 
 class ContainerCustomButton extends StatefulWidget {
   const ContainerCustomButton(
