+++ doce_conta/lib/widgets/conteiner_text_center.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import '../widgets/button.dart';
 
 class Conteinertextcenter extends StatefulWidget {
   const Conteinertextcenter({super.key, required this.btnText, this.onPressedFunction, required this.widhtContainer, required this.heightContainer});
